feat: dev tools button; open on new tab
