+++ src/dev/vite/devtools.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="/src/dev/lib/styles.css" />
+    <title>Dev tools</title>
+  </head>
+  <body>
+    <div id="devtools-root"></div>
+  </body>
+</html>
